@@ -1,25 +1,25 @@
-import axios from 'axios';
-import { clearStore } from '../features/user/userSlice';
-import { getUserFromLocalStorage } from './localStorage';
+import axios from 'axios'
+import { clearStore } from '../features/user/userSlice'
+import { getUserFromLocalStorage } from './localStorage'
 
 const customFetch = axios.create({
   baseURL: '/api/v1',
-});
+})
 
 customFetch.interceptors.request.use((config) => {
-  const user = getUserFromLocalStorage();
+  const user = getUserFromLocalStorage()
   if (user) {
-    config.headers.common['Authorization'] = `Bearer ${user.token}`;
+    config.headers['Authorization'] = `Bearer ${user.token}`
   }
-  return config;
-});
+  return config
+})
 
 export const checkForUnauthorizedResponse = (error, thunkAPI) => {
   if (error.response.status === 401) {
-    thunkAPI.dispatch(clearStore());
-    return thunkAPI.rejectWithValue('Unauthorized! Logging Out...');
+    thunkAPI.dispatch(clearStore())
+    return thunkAPI.rejectWithValue('Unauthorized! Logging Out...')
   }
-  return thunkAPI.rejectWithValue(error.response.data.msg);
-};
+  return thunkAPI.rejectWithValue(error.response.data.msg)
+}
 
-export default customFetch;
+export default customFetch
